feat: 기능별 파일 분리 및 cell 내의 product 프로퍼티 생성
