+++ OpenMarket/OpenMarket/MainViewController/CollectionView/Grid/GridCell.swift
@@ -2,12 +2,13 @@
 //  ListCell.swift
 //  OpenMarket
 //
-//  Created by Jpush, Aaron on 2022/11/22.
+//  Created by Jpush, Aaron on 2022/11/24.
 //
 
 import UIKit
 
-final class ListCell: UICollectionViewListCell {
+final class GridCell: UICollectionViewCell {
+    var product: Product?
     
     let image: UIImageView = {
         let imageView = UIImageView()
@@ -18,7 +19,7 @@ final class ListCell: UICollectionViewListCell {
     let productNameLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
-        label.textAlignment = .left
+        label.textAlignment = .center
         label.numberOfLines = 0
         label.font = UIFont.preferredFont(forTextStyle: .headline)
         label.adjustsFontForContentSizeCategory = true
@@ -34,6 +35,7 @@ final class ListCell: UICollectionViewListCell {
         label.attributedText = attributeString
         label.textColor = .systemRed
         
+        label.textAlignment = .center
         label.font = UIFont.preferredFont(forTextStyle: .subheadline)
         label.adjustsFontForContentSizeCategory = true
         return label
@@ -44,6 +46,7 @@ final class ListCell: UICollectionViewListCell {
         label.translatesAutoresizingMaskIntoConstraints = false
         label.font = UIFont.preferredFont(forTextStyle: .subheadline)
         label.textColor = .systemGray
+        label.textAlignment = .center
         label.adjustsFontForContentSizeCategory = true
         return label
     }()
@@ -51,70 +54,55 @@ final class ListCell: UICollectionViewListCell {
     let stockLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
-        label.textAlignment = .right
         label.numberOfLines = 0
+        label.textAlignment = .center
         label.font = UIFont.preferredFont(forTextStyle: .subheadline)
         label.adjustsFontForContentSizeCategory = true
         label.textColor = .systemGray
         return label
     }()
     
-    let nameStockStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.axis = .horizontal
-//        stackView.distribution = .fill
-        stackView.alignment = .center
-        return stackView
-    }()
-    
     let priceStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.axis = .horizontal
-        stackView.spacing = 8
-        return stackView
-    }()
-    
-    let labelStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.translatesAutoresizingMaskIntoConstraints = false
         stackView.axis = .vertical
-        stackView.distribution = .fill
-//        stackView.spacing = 8
+        stackView.alignment = .center
+
         return stackView
     }()
     
     let containerStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.axis = .horizontal
+        stackView.axis = .vertical
         stackView.alignment = .center
-        stackView.distribution = .fill
-        stackView.spacing = 8
+        stackView.distribution = .equalSpacing
+        
         return stackView
     }()
     
     override init(frame: CGRect) {
         super.init(frame: .zero)
         
-        self.accessories = [.disclosureIndicator()]
-        
         self.contentView.addSubview(containerStackView)
         
         self.containerStackView.addArrangedSubview(image)
-        self.containerStackView.addArrangedSubview(labelStackView)
+    
+        self.containerStackView.addArrangedSubview(productNameLabel)
+        self.contentView.layer.cornerRadius = 10
+        self.contentView.layer.borderWidth = 2
         
-        self.labelStackView.addArrangedSubview(nameStockStackView)
-        self.labelStackView.addArrangedSubview(priceStackView)
-        
-        self.nameStockStackView.addArrangedSubview(productNameLabel)
-        self.nameStockStackView.addArrangedSubview(stockLabel)
+        self.containerStackView.addArrangedSubview(priceStackView)
         
         self.priceStackView.addArrangedSubview(priceLabel)
         self.priceStackView.addArrangedSubview(discountedPriceLabel)
         
+        self.containerStackView.addArrangedSubview(stockLabel)
         setUpUI()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
     }
     
     override func prepareForReuse() {
@@ -127,39 +115,17 @@ final class ListCell: UICollectionViewListCell {
         priceLabel.isHidden = false
     }
     
-    func setUpUI() {
-        priceLabel.setContentHuggingPriority(.defaultHigh - 1, for: .horizontal)
-        
+    private func setUpUI() {
         NSLayoutConstraint.activate([
-            containerStackView.topAnchor.constraint(equalTo: contentView.topAnchor),
-            containerStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            containerStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
-            containerStackView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
+            containerStackView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 12),
+            containerStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 12),
+            containerStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -12),
+            containerStackView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -12),
             
-            image.widthAnchor.constraint(equalTo: contentView.widthAnchor, multiplier: 1/6),
+            image.widthAnchor.constraint(equalTo: containerStackView.widthAnchor),
             image.heightAnchor.constraint(equalTo: image.widthAnchor),
-            
-//            stock.widthAnchor.constraint(lessThanOrEqualTo: nameStockStackView.widthAnchor, multiplier: 0.4),
-//
-//            nameStockStackView.heightAnchor.constraint(greaterThanOrEqualTo: productName.heightAnchor)
-//            stock.heightAnchor.constraint(greaterThanOrEqualTo: productName.heightAnchor),
-//            productName.widthAnchor.constraint(equalTo: contentView.widthAnchor, multiplier: 0.6)
-//            contentView.heightAnchor.constraint(greaterThanOrEqualTo: image.heightAnchor)
-//            labelStackView.heightAnchor.constraint(greaterThanOrEqualTo: image.heightAnchor)
-            
         ])
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
     }
 }
 
-extension ListCell {
-    private func createStrikethroughAttribute() -> NSMutableAttributedString {
-        let attributeString: NSMutableAttributedString = NSMutableAttributedString(string: " ")
-        attributeString.addAttribute(NSAttributedString.Key.strikethroughStyle, value: 1, range: NSRange(location: 0, length: attributeString.length))
-        
-        return attributeString
-    }
-}
+
