+++ OpenMarket/OpenMarket/MainViewController/CollectionView/List/ListCell.swift
@@ -2,12 +2,14 @@
 //  ListCell.swift
 //  OpenMarket
 //
-//  Created by Jpush, Aaron on 2022/11/24.
+//  Created by Jpush, Aaron on 2022/11/22.
 //
 
 import UIKit
 
-final class GridCell: UICollectionViewCell {
+final class ListCell: UICollectionViewListCell {
+    var product: Product?
+    
     let image: UIImageView = {
         let imageView = UIImageView()
         imageView.translatesAutoresizingMaskIntoConstraints = false
@@ -17,8 +19,8 @@ final class GridCell: UICollectionViewCell {
     let productNameLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
+        label.textAlignment = .left
         label.numberOfLines = 0
-        label.textAlignment = .center
         label.font = UIFont.preferredFont(forTextStyle: .headline)
         label.adjustsFontForContentSizeCategory = true
         return label
@@ -33,7 +35,6 @@ final class GridCell: UICollectionViewCell {
         label.attributedText = attributeString
         label.textColor = .systemRed
         
-        label.textAlignment = .center
         label.font = UIFont.preferredFont(forTextStyle: .subheadline)
         label.adjustsFontForContentSizeCategory = true
         return label
@@ -44,7 +45,6 @@ final class GridCell: UICollectionViewCell {
         label.translatesAutoresizingMaskIntoConstraints = false
         label.font = UIFont.preferredFont(forTextStyle: .subheadline)
         label.textColor = .systemGray
-        label.textAlignment = .center
         label.adjustsFontForContentSizeCategory = true
         return label
     }()
@@ -52,55 +52,70 @@ final class GridCell: UICollectionViewCell {
     let stockLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
+        label.textAlignment = .right
         label.numberOfLines = 0
-        label.textAlignment = .center
         label.font = UIFont.preferredFont(forTextStyle: .subheadline)
         label.adjustsFontForContentSizeCategory = true
         label.textColor = .systemGray
         return label
     }()
     
+    let nameStockStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        stackView.axis = .horizontal
+//        stackView.distribution = .fill
+        stackView.alignment = .center
+        return stackView
+    }()
+    
     let priceStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
+        stackView.axis = .horizontal
+        stackView.spacing = 8
+        return stackView
+    }()
+    
+    let labelStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.translatesAutoresizingMaskIntoConstraints = false
         stackView.axis = .vertical
-        stackView.alignment = .center
-
+        stackView.distribution = .fill
+//        stackView.spacing = 8
         return stackView
     }()
     
     let containerStackView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.axis = .vertical
+        stackView.axis = .horizontal
         stackView.alignment = .center
-        stackView.distribution = .equalSpacing
-        
+        stackView.distribution = .fill
+        stackView.spacing = 8
         return stackView
     }()
     
     override init(frame: CGRect) {
         super.init(frame: .zero)
         
+        self.accessories = [.disclosureIndicator()]
+        
         self.contentView.addSubview(containerStackView)
         
         self.containerStackView.addArrangedSubview(image)
-    
-        self.containerStackView.addArrangedSubview(productNameLabel)
-        self.contentView.layer.cornerRadius = 10
-        self.contentView.layer.borderWidth = 2
+        self.containerStackView.addArrangedSubview(labelStackView)
         
-        self.containerStackView.addArrangedSubview(priceStackView)
+        self.labelStackView.addArrangedSubview(nameStockStackView)
+        self.labelStackView.addArrangedSubview(priceStackView)
+        
+        self.nameStockStackView.addArrangedSubview(productNameLabel)
+        self.nameStockStackView.addArrangedSubview(stockLabel)
         
         self.priceStackView.addArrangedSubview(priceLabel)
         self.priceStackView.addArrangedSubview(discountedPriceLabel)
         
-        self.containerStackView.addArrangedSubview(stockLabel)
         setUpUI()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
     }
     
     override func prepareForReuse() {
@@ -113,17 +128,39 @@ final class GridCell: UICollectionViewCell {
         priceLabel.isHidden = false
     }
     
-    private func setUpUI() {
+    func setUpUI() {
+        priceLabel.setContentHuggingPriority(.defaultHigh - 1, for: .horizontal)
+        
         NSLayoutConstraint.activate([
-            containerStackView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 12),
-            containerStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 12),
-            containerStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -12),
-            containerStackView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -12),
+            containerStackView.topAnchor.constraint(equalTo: contentView.topAnchor),
+            containerStackView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
+            containerStackView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
+            containerStackView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor),
             
-            image.widthAnchor.constraint(equalTo: containerStackView.widthAnchor),
+            image.widthAnchor.constraint(equalTo: contentView.widthAnchor, multiplier: 1/6),
             image.heightAnchor.constraint(equalTo: image.widthAnchor),
+            
+//            stock.widthAnchor.constraint(lessThanOrEqualTo: nameStockStackView.widthAnchor, multiplier: 0.4),
+//
+//            nameStockStackView.heightAnchor.constraint(greaterThanOrEqualTo: productName.heightAnchor)
+//            stock.heightAnchor.constraint(greaterThanOrEqualTo: productName.heightAnchor),
+//            productName.widthAnchor.constraint(equalTo: contentView.widthAnchor, multiplier: 0.6)
+//            contentView.heightAnchor.constraint(greaterThanOrEqualTo: image.heightAnchor)
+//            labelStackView.heightAnchor.constraint(greaterThanOrEqualTo: image.heightAnchor)
+            
         ])
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
     }
 }
 
-
+extension ListCell {
+    private func createStrikethroughAttribute() -> NSMutableAttributedString {
+        let attributeString: NSMutableAttributedString = NSMutableAttributedString(string: " ")
+        attributeString.addAttribute(NSAttributedString.Key.strikethroughStyle, value: 1, range: NSRange(location: 0, length: attributeString.length))
+        
+        return attributeString
+    }
+}
