+++ OpenMarket/OpenMarket/ProductViewController/UIImage+Extension.swift
@@ -1,0 +1,19 @@
+//
+//  UIImage+Extension.swift
+//  OpenMarket
+//
+//  Created by 이정민 on 2022/12/05.
+//
+
+import UIKit
+
+extension UIImage {
+    func resizeImageTo(size: CGSize) -> UIImage {
+        UIGraphicsBeginImageContextWithOptions(size, false, 0.0)
+        self.draw(in: CGRect(origin: CGPoint.zero, size: size))
+        let resizedImage = UIGraphicsGetImageFromCurrentImageContext()!
+        UIGraphicsEndImageContext()
+        
+        return resizedImage
+    }
+}
